edit host
