+++ src/api/index.js
@@ -1205,3 +1205,7 @@ export function queryServiceWorkloads (projectName, envName) {
 export function postWorkloads (payload) {
   return http.post(`/api/aslan/service/workloads`, payload)
 }
+
+export function editWorkloads (payload) {
+  return http.put(`/api/aslan/service/workloads`, payload)
+}
